+++ sdk/monitor/monitor-opentelemetry-exporter/src/types.ts
@@ -190,6 +190,7 @@ export enum experimentalOpenTelemetryValues {
   SYNTHETIC_TYPE = "user_agent.synthetic.type",
   ATTR_ENDUSER_PSEUDO_ID = "enduser.pseudo.id",
   ATTR_ENDUSER_ID = "enduser.id",
+  ATTR_SESSION_ID = "session.id",
 }
 
 /**
@@ -221,6 +222,7 @@ export const httpSemanticValues = [
   ATTR_EXCEPTION_STACKTRACE,
   experimentalOpenTelemetryValues.ATTR_ENDUSER_ID,
   experimentalOpenTelemetryValues.ATTR_ENDUSER_PSEUDO_ID,
+  experimentalOpenTelemetryValues.ATTR_SESSION_ID,
   experimentalOpenTelemetryValues.SYNTHETIC_TYPE,
 ];